requirements.txt
supervisor
fastapi[standard]
aiohttp
python-dotenv
